// apzkr-pzpi-21-3-merkulov-arsenii/Task3-WebClient/hospitality-hub/src/app/api-proxy/models/room-response.ts
/* tslint:disable */
/* eslint-disable */
import { ERoomStatus } from '../models/e-room-status';
import { ERoomType } from '../models/e-room-type';
export interface RoomResponse {
  apiKey?: string | null;
  baseLockUri?: string | null;
  basePrice?: number;
  discountPercent?: number;
  hotelId?: number;
  id?: number;
  number?: number;
  roomType?: ERoomType;
  stage?: number;
  status?: ERoomStatus;
}